+++ CHANGELOG.md
@@ -5,7 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/nuwave/lighthouse/compare/v4.5.2...master)
+## [Unreleased](https://github.com/nuwave/lighthouse/compare/v4.5.3...master)
+
+### Added
+
+- Add `@scope` directive for adding a scope to the query builder https://github.com/nuwave/lighthouse/pull/998
+
+## [4.5.3](https://github.com/nuwave/lighthouse/compare/v4.5.2...v4.5.3)
+
+### Fixed
+
+- Handle `null` being passed to a nullable argument that is an input object type https://github.com/nuwave/lighthouse/pull/1021
 
 ## [4.5.2](https://github.com/nuwave/lighthouse/compare/v4.5.1...v4.5.2)